FoodDelivery: add cart implementation

+++ src/main/java/com/cogent/fooddeliveryapp/controller/CustomerController.java
@@ -6,9 +6,14 @@ import java.util.Optional;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -18,17 +23,25 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.cogent.fooddeliveryapp.dto.Address;
+import com.cogent.fooddeliveryapp.dto.CartItem;
 import com.cogent.fooddeliveryapp.dto.Customer;
+import com.cogent.fooddeliveryapp.dto.CustomerCart;
+import com.cogent.fooddeliveryapp.dto.Food;
 import com.cogent.fooddeliveryapp.dto.Role;
 import com.cogent.fooddeliveryapp.enums.UserRoles;
 import com.cogent.fooddeliveryapp.exceptions.CustomerNotFoundException;
+import com.cogent.fooddeliveryapp.exceptions.FoodNotFoundException;
 import com.cogent.fooddeliveryapp.exceptions.InvalidRequestException;
 import com.cogent.fooddeliveryapp.exceptions.NoRecordsFoundException;
 import com.cogent.fooddeliveryapp.exceptions.RoleNotFoundException;
 import com.cogent.fooddeliveryapp.payload.request.AddressRequest;
+import com.cogent.fooddeliveryapp.payload.request.CartUpdateRequest;
 import com.cogent.fooddeliveryapp.payload.request.CustomerRegistrationRequest;
+import com.cogent.fooddeliveryapp.payload.response.CartStatusResponse;
 import com.cogent.fooddeliveryapp.payload.response.CustomerResponse;
+import com.cogent.fooddeliveryapp.security.services.UserDetailsImpl;
 import com.cogent.fooddeliveryapp.service.CustomerService;
+import com.cogent.fooddeliveryapp.service.FoodService;
 import com.cogent.fooddeliveryapp.service.RoleService;
 
 /**
@@ -47,13 +60,31 @@ public class CustomerController {
 	@Autowired
 	private Function<AddressRequest, Address> addressMapper;
 
+	@Autowired
+	private FoodService foodService;
+	
+	private Authentication getAuthentication() {
+		return SecurityContextHolder.getContext().getAuthentication();
+	}
+	
+	// Checks if user has access to this item
+	// Only ADMINs or the specified user can access
+	private void checkUserAccess(int userID) {
+		Authentication auth = getAuthentication();
+		UserDetailsImpl userDetails = (UserDetailsImpl) auth.getPrincipal();
+		
+		if (!userDetails.isAdmin() && userDetails.getId() != userID) {
+			throw new AccessDeniedException("User unable to access this resource");
+		}
+	}
+
 	@GetMapping
 	@PreAuthorize("hasRole('ROLE_ADMIN')")
 	public ResponseEntity<?> getUsers() throws NoRecordsFoundException {
 		List<Customer> customers = customerService.getAllCustomers();
 		
 		if (customers != null && !customers.isEmpty()) {
-			return ResponseEntity.ok(customers.parallelStream().map(user -> {
+			return ResponseEntity.ok(customers.stream().map(user -> {
 				return new CustomerResponse(user);
 			}).sorted((o1, o2) -> {
 				return Integer.compare(o1.getId(), o2.getId());
@@ -64,7 +95,10 @@ public class CustomerController {
 	}
 	
 	@GetMapping("/{userID}")
+	@PreAuthorize("hasRole('ROLE_ADMIN') || hasRole('ROLE_USER')")
 	public ResponseEntity<?> getUserByID(@PathVariable int userID) throws NoRecordsFoundException {
+		checkUserAccess(userID);
+		
 		Optional<Customer> customerOpt = customerService.getCustomerByID(userID);
 		if (customerOpt.isPresent()) {
 			return ResponseEntity.ok(new CustomerResponse(customerOpt.get()));
@@ -74,7 +108,10 @@ public class CustomerController {
 	}
 	
 	@PutMapping("/{userID}")
-	public ResponseEntity<?> updateUserByID(@PathVariable int userID, @RequestBody CustomerRegistrationRequest request) throws NoRecordsFoundException, CustomerNotFoundException, InvalidRequestException {
+	@PreAuthorize("hasRole('ROLE_ADMIN') || hasRole('ROLE_USER')")
+	public ResponseEntity<?> updateUserByID(@PathVariable int userID, @Valid @RequestBody CustomerRegistrationRequest request) throws NoRecordsFoundException, CustomerNotFoundException, InvalidRequestException {
+		checkUserAccess(userID);
+		
 		Optional<Customer> customerOpt = customerService.getCustomerByID(userID);
 		if (customerOpt.isPresent()) {
 			final Customer customer = customerOpt.get();
@@ -88,7 +125,7 @@ public class CustomerController {
 			customer.setPassword(request.getPassword());
 			customer.setName(request.getName());
 			customer.setDoj(request.getDoj());		
-			customer.setAddresses(request.getAddress().parallelStream().map(addressRequest -> {
+			customer.setAddresses(request.getAddress().stream().map(addressRequest -> {
 				Address address = addressMapper.apply(addressRequest);
 				address.setCustomer(customer);
 				return address;
@@ -122,7 +159,10 @@ public class CustomerController {
 	}
 	
 	@DeleteMapping("/{userID}")
+	@PreAuthorize("hasRole('ROLE_ADMIN') || hasRole('ROLE_USER')")
 	public ResponseEntity<?> deleteUserByID(@PathVariable int userID) throws NoRecordsFoundException {
+		checkUserAccess(userID);
+
 		boolean exists = customerService.existsByID(userID);
 		if (exists) {
 			customerService.deleteCustomerByID(userID);
@@ -131,5 +171,66 @@ public class CustomerController {
 		} else {
 			throw new NoRecordsFoundException("Customer with ID: " + userID + " not found");
 		}
+	}
+	
+	/* Customer Cart operations */
+	@GetMapping("/{userID}/cart")
+	@PreAuthorize("hasRole('ROLE_ADMIN') || hasRole('ROLE_USER')")
+	public ResponseEntity<?> getUserCartByID(@PathVariable int userID) throws NoRecordsFoundException, CustomerNotFoundException {
+		checkUserAccess(userID);
+		
+		Customer customer = customerService.getCustomerByID(userID).orElseThrow(() -> {
+			return new NoRecordsFoundException("Customer with ID: " + userID + " not found");
+		});
+		
+		return ResponseEntity.ok(new CartStatusResponse(customer.getCustomerCart()));
+	}
+	
+	@PutMapping("/{userID}/cart")
+	@PreAuthorize("hasRole('ROLE_ADMIN') || hasRole('ROLE_USER')")
+	public ResponseEntity<?> updateUserCartByID(@PathVariable int userID, @Valid @RequestBody CartUpdateRequest request) throws NoRecordsFoundException, CustomerNotFoundException {
+		checkUserAccess(userID);
+		
+		Customer customer = customerService.getCustomerByID(userID).orElseThrow(() -> {
+			return new NoRecordsFoundException("Customer with ID: " + userID + " not found");
+		});
+		
+		CustomerCart cart = customer.getCustomerCart();
+		cart.getCartItems().clear();
+		cart.getCartItems().addAll(request.getCart().stream().map(foodID -> {
+			Food foodItem = foodService.getFoodByID(foodID).orElseThrow(() -> {
+				return new FoodNotFoundException("Food with ID: " + foodID + " not found");
+			});
+			
+			return new CartItem(foodItem, cart);
+		}).collect(Collectors.toSet()));
+		cart.setActive(true);
+		
+		// Update customer
+		customer = customerService.updateCustomer(customer);
+
+		return ResponseEntity.ok(new CartStatusResponse(customer.getCustomerCart()));
+	}
+	
+	@PutMapping("/{userID}/cart/checkout")
+	@PreAuthorize("hasRole('ROLE_ADMIN') || hasRole('ROLE_USER')")
+	public ResponseEntity<?> checkoutUserCartByID(@PathVariable int userID) throws NoRecordsFoundException, CustomerNotFoundException {
+		checkUserAccess(userID);
+		
+		Customer customer = customerService.getCustomerByID(userID).orElseThrow(() -> {
+			return new NoRecordsFoundException("Customer with ID: " + userID + " not found");
+		});
+		
+		CustomerCart cart = customer.getCustomerCart();
+		System.out.println("Pre checkout cart items: " + cart.getCartItems());
+		cart.getCartItems().clear();
+		cart.setActive(false);
+		System.out.println("Post checkout cart items: " + cart.getCartItems());
+		
+		// Update customer
+		customer = customerService.updateCustomer(customer);
+		System.out.println("Post update customer: " + customer.getCustomerCart().getCartItems());
+
+		return ResponseEntity.ok(new CartStatusResponse(customer.getCustomerCart()));
 	}
 }
